org-media-note-core: handle org-ref notes files without a CUSTOM_ID
bibtex-completions.el (which is used by org-ref) presently allows
bibtex-completion-notes-path to be a directory. the doc string for the
variable reads.

  "The place where notes are stored.  This is either a file, in which
  case all notes are stored in that file, or a directory, in which
  case each publication gets its own notes file in that directory.  In
  the latter case, bibtex-completion assumes that the names of the
  note files are composed of the BibTeX-key plus a suffix that is
  specified in `bibtex-completion-notes-extension'."

In this case (when it is a directory) there is no CUSTOM_ID to get the
org-ref key.

* org-media-note-core.el:
(org-media-note--org-ref-key-from-bibtex-completion-notes): get the
org ref key from the file name if the notes file is in a notes
directory managed by
bibtex-complete.el.  (org-media-note--current-org-ref-key): use it
here.
* README.org: update.