org-media-note-core: handle media files from org-ref links
Again no need to insist on CUSTOM_ID to get the org-ref link.  If you
are on a org-ref citation and using org-media-note-use-org-ref, it
should be possible to play the related media file directly.

* org-media-note-org-ref.el: (org-media-note--org-ref-key-from-cite):
return the key if you are on an org-ref citation link.

* org-media-note-core.el: (org-media-note--current-org-ref-key): if
using org-ref use above function (in org-media-note-org-ref.el) to get
the key. (Yes this is a boundary abstraction violation)

[230423] refactor
* org-media-note-org-ref.el: (org-media-note-link-types): new variable
to hold the "videocite" and "audiocite" types.
(org-media-note-media-cite-link-message): refactor slightly to use
org-media-note-link-types.

[230625]
* org-media-note-org-ref.el: (org-media-note-setup-org-ref): handle
audiocite and videocite. handle timestamps in audiocite and videocite.

[240307]
* org-media-note-core.el: (org-media-note--current-org-ref-key) -
don't try to use org-media-note-ref-cite-p over org-media-note-use-org-ref.
not just because it is expensive but because it will also lead to an infinte loop

[230625]
org-media-note--current-org-ref-key: (org-media-note--current-org-ref-key):

fixes commit
"org-media-note-core: handle media files from org-ref links"