org-media-note-core: handle media files from org-ref links
Again no need to insist on CUSTOM_ID to get the org-ref link.  If you
are on a org-ref citation and using org-media-note-use-org-ref, it
should be possible to play the related media file directly.

* org-media-note-org-ref.el: (org-media-note--org-ref-key-from-cite):
return the key if you are on an org-ref citation link.

* org-media-note-core.el: (org-media-note--current-org-ref-key): if
using org-ref use above function (in org-media-note-org-ref.el) to get
the key. (Yes this is a boundary abstraction violation)