org-media-note: new action: Update Link Timestamp
"Adjust timestamp" doesn't quite do what I want it to do: which is
update the timestamp in the citation link to what is currently
playing. So the new function.

* org-media-note-core.el: (org-media-note--link-extract-link): extract
the link portion (not desc) from the string produced
by (org-media-note--link). (org-media-note-update-link): implement.

[240208]
tweak crude cl-search matcher to match "]]" in addition to "][".

[241017]
* org-media-note.el: (org-media-note-transient): bind "u"