org-media-note: new action: Update Link Timestamp
TODO: Update README.

"Adjust timestamp" doesn't quite do what I want it to do: which is
update the timestamp in the org-ref cite link to what is currently
playing. So the new function.

* contrib/omn-{hydra,transient}.el, org-media-note-pretty-hydra.el: add
keybinding "u" to the hydra and transient dialogs

* org-media-note-core.el: (org-media-note--link-extract-link): extract
the link portion (not desc) from the string produced
by (org-media-note--link). (org-media-note-update-link): implement.

[240208]
tweak crude cl-search matcher to match "]]" in addition to "][".

[241017]
* org-media-note-ui-transient.el: bind "u"