org-media-note-core.el: prefer refcite if currently playing file is known to bibtex-completion
* org-media-note-core.el (org-media-note-lookup-citation-for-file):
new. requires bibtex-completion-utils to get a function to map known
entries and use it to get a cite for the currently playing file. NOTE
the "file" key in the bib may be enclosed in {} bracket chars, we only
do a substring match via cl-search.

(org-media-note--current-citation-key): extend to return the currently
playing media file if it is known to bibtex-completion.

(org-media-note-ref-cite-p): TODO FIXME. revisit all callers.  We
should perhaps be caching the call
to (org-media-note--current-citation-key), if it immediately reused.

[240318]
* org-media-note-core.el: (org-media-note--link-base-file): assume
org-ref-cite-insert-version == 3 and use the "&cite" syntax.
Inspecting the variable would require requiring org-ref.

This way version 3 org-ref cites which are activated stay activated,
say after adjusting or updating the time stamp.

[240429]
* org-media-note-core.el: (org-media-note-lookup-citation-for-file):
handle NULL file