org-media-note-core.el: avoid calls to org-media-note-ref-cite-p
* org-media-note-core.el: (org-media-note--link,
org-media-note--link-base-file): avoid calling the function
`org-media-note-ref-cite-p'.  We want to insert a (refcite) link for
the current playing media (1), or lookup its file-path (2). we don't
care about the "current" media link at point in our org file.  so only
use the global variable `org-media-note-use-refcite-first' to decide
if a "refcite" is to be inserted (1) or used for lookup (2).  in both
cases we don't need to call the dubiously named function
`org-media-note-ref-cite-p', so don't call it.