avoid (org-media-note--ab-loop-p) in the various UI `backends'
using (org-media-note--ab-loop-p) does not update the UI after hitting
`l' twice (i.e after setting a and b), and shows the state
incorrectly: although both a and b are set, the ui thinks b is not set
yet, and hitting `l' would clear the a-b loop instead of setting b.

instead just checking if time-a and time-b are set instead
of (org-media-loop-ab-loop-p) to determine the state seems to help

* org-media-note-core.el: (org-media-note--ui-ab-loop-title): avoid
calling org-media-note--ab-loop-p.