org-media-note-cite.el: set activate-func for "audiocite" and "videocite" to org-ref-activate
for this to work correctly one has to patch org-ref to

- modify org-ref-citation-key-re in org-ref/org-ref-citation-links.el to remove
  `#' from the list of allowed characters in the key and label because
  of the syntax for timestamps.

for Link Activation for version 3 ref-cites. For version 2

-  modify org-ref-cite-activate in
in the version 2 branch to set
```
	 (setq substrings (cl-loop for key in substrings for p =
(cl-position ?\# key) if p collect (cl-subseq key 0 p) else collect p))
```

[240308]
* org-media-note-org-ref.el:
(org-media-note-working-with-kluged-org-ref-p): check if we're working
with a modified org-ref so the activate-funcs work. if so,
(org-media-note-ref-parse-path): put back the "#timestamp" into the
path, this would be found in the suffix.

[240318]
unconditionally modify the value of org-ref-citation-key-re  so org-ref
activation will treat `#' as part of the cite.

for version 2 ref cites we add an around advice.
(org-media-note-ref--org-ref-arse-cite-path--around-advice): around
org-ref-parse-cite-path so that function can drop the #fragments.

(org-media-note-working-with-kluged-org-ref-p) becomes trivial and
should be eliminated.

(org-media-note-ref-parse-path): take pains to reintroduce the
fragment, which is needed by org-media-note-ref functions, for both
versions.

[250704]
* org-media-note-cite.el: (org-media-note-cite-setup)
Initialize conditional on org-media-note-use-org-ref