org-media-note-cite.el: advice org-ref-find-bibliography in org-src buffers
* org-media-note-org-ref.el:
(org-ref-find-bibliography-in-org-src-buffer-around-advice): set
current-buffer to the original org file when calling
`org-ref-find-bibliography'

org-ref bib related functions that depend on org-ref-find-bibliography
do not work in org-src buffers (which is not an org-mode buffer) so
wrap calls to this function by setting the current buffer to the
original org buffer.

also note: bibtex-completion-bibliography is buffer-local to the org
file, and will not be available for org-ref bib related functions.